+++ codex-rs/core/src/lib.rs
@@ -15,7 +15,6 @@ pub mod codex;
 mod codex_conversation;
 pub mod token_data;
 pub use codex_conversation::CodexConversation;
-pub mod approvals;
 mod command_safety;
 pub mod config;
 pub mod config_edit;
@@ -38,10 +37,7 @@ mod mcp_connection_manager;
 mod mcp_tool_call;
 mod message_history;
 mod model_provider_info;
-pub mod orchestrator;
 pub mod parse_command;
-mod pty;
-mod runtime;
 pub mod sandboxing;
 mod truncate;
 mod unified_exec;
@@ -63,7 +59,6 @@ pub use auth::CodexAuth;
 pub mod default_client;
 pub mod model_family;
 mod openai_model_info;
-mod openai_tools;
 pub mod project_doc;
 mod rollout;
 pub(crate) mod safety;
