--- codex-rs/core/src/openai_tools.rs
@@ -1,1 +1,0 @@
-pub use crate::tools::spec::*;
