+++ codex-rs/core/src/tools/runtimes/mod.rs
@@ -4,6 +4,6 @@ Module: runtimes
 Concrete ToolRuntime implementations for specific tools. Each runtime stays
 small and focused and reuses the orchestrator for approvals + sandbox + retry.
 */
-pub mod apply_patch_runtime;
+pub mod apply_patch;
 pub mod command_spec;
-pub mod shell_runtime;
+pub mod shell;
